+++ packages/experimental/AppearanceAdditions/src/appearanceAdditions.ts
@@ -1,0 +1,7 @@
+import { AppearanceAdditions } from './NativeAppearanceAdditions.types';
+
+// Default values for non-iOS clients.
+export const appearanceAdditions = {
+  horizontalSizeClass: 'regular',
+  userInterfaceLevel: 'base',
+} as AppearanceAdditions;