+++ packages/experimental/AppearanceAdditions/src/getSizeClass.ts
@@ -1,0 +1,6 @@
+import { SizeClass } from './NativeAppearanceAdditions.types';
+
+export function useHorizontalSizeClass(): SizeClass {
+  // Stubbed out for non-iOS platforms
+  return 'regular';
+}