+++ packages/experimental/Shimmer/src/Shimmer.styling.win32.ts
@@ -1,19 +1,7 @@
-import { UseStylingOptions, buildProps, Theme, TokenSettings } from '@fluentui-react-native/framework';
+import { UseStylingOptions, buildProps } from '@fluentui-react-native/framework';
 import { shimmerName, ShimmerProps, ShimmerTokens } from './Shimmer.types';
 import { ShimmerSlotProps } from './Shimmer.types.win32';
-
-export const defaultShimmerTokens: TokenSettings<ShimmerTokens> = (theme: Theme) =>
-  ({
-    angle: 0,
-    backgroundColor: theme.colors.background,
-    delay: 500,
-    duration: 2000,
-    shimmerColor: theme.colors.bodyFrameDivider,
-    shimmerColorOpacity: 1,
-    shimmerWaveColor: '#E1E1E1',
-    shimmerWaveColorOpacity: 1,
-    shimmerWaveWidth: '100%',
-  } as ShimmerTokens);
+import { defaultShimmerTokens } from './ShimmerTokens.win32';
 
 /**
  * tell the styling hook how to build up the tokens