+++ packages/experimental/MenuButton/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui-react-native/experimental-menu-button
 
-This log was last generated on Fri, 20 Jan 2023 21:49:21 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Jan 2023 21:31:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.22
+
+Wed, 25 Jan 2023 21:31:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/contextual-menu to v0.21.21
+- Bump @fluentui-react-native/experimental-button to v0.16.83
+- Bump @fluentui-react-native/framework to v0.8.36
+- Bump @fluentui-react-native/tokens to v0.20.5
 
 ## 0.6.21