+++ packages/theming/theming-utils/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui-react-native/theming-utils
 
-This log was last generated on Fri, 20 Jan 2023 18:06:46 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Jan 2023 21:31:17 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.23.0
+
+Wed, 25 Jan 2023 21:31:17 GMT
+
+### Minor changes
+
+- adding shared colors android (email not defined)
+- Bump @fluentui-react-native/theme-types to v0.30.0
+- Bump @fluentui-react-native/tokens to v0.20.5
 
 ## 0.22.2